requirements.txt
certifi==2021.5.30
chardet==4.0.0
Click==8.0.1
Flask~=1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
requests==2.26.0
stripe==2.60.0
toml==0.10.2
Werkzeug==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
python-dotenv~=0.19.0